pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "AI Security Threat Intelligence Agent for Guard0"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.43.0",
    "feedparser>=6.0.11",
    "httpx>=0.27.0",
    "slack-sdk>=3.33.0",
    "twilio>=9.0.0",
    "click>=8.1.7",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
argus = "argus.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/argus"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
